src/requirements.txt
langchain
langchain-community
openai
slack_sdk
notion-util
blockkit
slack_bolt
notion-client
PyGithub
tiktoken
requests
selenium
graphviz
gunicorn
flask
pluginbase
google-api-python-client
youtube-transcript-api
functions-framework
sentry-sdk
PyPDF2==2.12.1
google-generativeai
pillow
beautifulsoup4
pytz
yara-python
plyara
boto3
awslambdaric
